backend/requirements.txt
fastapi==0.115.3
zyte-api==0.6.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3